+++ zest/releaser/postrelease.py
@@ -25,6 +25,7 @@ DATA = {
     # in self.data that is not in this list.  Embarrasment-driven
     # documentation!
     'workingdir': 'Original working directory',
+    'reporoot': 'Root of the version control repository',
     'name': 'Name of the project being released',
     'nothing_changed_yet': 'First line in new changelog section',
     'new_version': 'New development version (so 1.1)',